rename modal context provider
