+++ frontend/src/components/modals/RegisterModal.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Form, Button } from "reactstrap";
 import BaseModal from "../base/BaseModal";
-import { useModal } from "../../contexts/ModalContextProvider";
+import { useModal } from "../../contexts/ModalContext";
 
 const RegisterModal = () => {
